+++ stacker/actions/diff.py
@@ -3,6 +3,8 @@ import json
 import logging
 from operator import attrgetter
 
+import yaml
+
 from . import build
 from .. import exceptions
 from ..plan import COMPLETE, Plan
@@ -143,7 +145,8 @@ def print_stack_changes(stack_name, new_stack, old_stack, new_params,
     to_file = "new_%s" % (stack_name,)
     lines = difflib.context_diff(
         old_stack, new_stack,
-        fromfile=from_file, tofile=to_file)
+        fromfile=from_file, tofile=to_file,
+        n=7)  # ensure at least a few lines of context are displayed afterward
 
     template_changes = list(lines)
     if not template_changes:
@@ -211,7 +214,7 @@ class Action(build.Action):
 
         stack.resolve(self.context, self.provider)
         # generate our own template & params
-        new_template = stack.blueprint.rendered
+        new_template = stack.blueprint.to_json()
         parameters = self.build_parameters(stack)
         new_params = dict()
         for p in parameters:
@@ -224,7 +227,18 @@ class Action(build.Action):
             self._print_new_stack(new_stack, parameters)
         else:
             # Diff our old & new stack/parameters
-            old_stack = self._normalize_json(old_template)
+            old_template = yaml.load(old_template)
+            if isinstance(old_template, str):
+                # YAML templates returned from CFN need parsing again
+                # "AWSTemplateFormatVersion: \"2010-09-09\"\nParam..."
+                # ->
+                # AWSTemplateFormatVersion: "2010-09-09"
+                old_template = yaml.load(old_template)
+            old_stack = self._normalize_json(
+                json.dumps(old_template,
+                           sort_keys=True,
+                           indent=4)
+            )
             print_stack_changes(stack.name, new_stack, old_stack, new_params,
                                 old_params)
         return COMPLETE